+++ src/main/java/com/acme/yupanaapi/domain/repository/HistorialRepository.java
@@ -4,8 +4,7 @@ import com.acme.yupanaapi.domain.model.Historial;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
-public interface HistorialRepository extends JpaRepository<Historial,Long> {
-    List<Historial> findBySellerId(Long sellerId);
+public interface HistorialRepository extends JpaRepository<Historial,Integer> {
+    List<Historial> findBySellerId(int sellerId);
 }
