+++ src/main/java/com/acme/yupanaapi/domain/repository/SaleRepository.java
@@ -4,6 +4,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.acme.yupanaapi.domain.model.Sale;
 
-public interface SaleRepository extends JpaRepository<Sale, Long>{
+public interface SaleRepository extends JpaRepository<Sale, Integer>{
 	List<Sale> findAllBySaleDate(Date saleDate);
 }
