+++ src/main/java/com/acme/yupanaapi/domain/repository/SellerRepository.java
@@ -3,7 +3,7 @@ import com.acme.yupanaapi.domain.model.Seller;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface SellerRepository extends JpaRepository<Seller,Long> {
-    Optional<Seller> findByIdAndUserId(Long id, Long userId);
+public interface SellerRepository extends JpaRepository<Seller,Integer> {
+    Optional<Seller> findByIdAndUserId(int id, int userId);
     Optional<Seller> findByBusinessName(String businessName);
 }
