+++ src/main/java/com/acme/yupanaapi/domain/repository/SubscriptionRepository.java
@@ -4,7 +4,7 @@ import com.acme.yupanaapi.domain.model.Subscription;
 
 import java.util.List;
 
-public interface SubscriptionRepository extends JpaRepository<Subscription, Long>{
-    List<Subscription> findAllByWalletId(Long walletId);
+public interface SubscriptionRepository extends JpaRepository<Subscription, Integer>{
+    List<Subscription> findAllByWalletId(int walletId);
 
 }
