+++ src/main/java/com/acme/yupanaapi/domain/repository/TransactionRepository.java
@@ -3,7 +3,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.acme.yupanaapi.domain.model.Transaction;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Long>{
-	List<Transaction> findAllByFlowId(Long flowId);
-	List<Transaction> findAllByHistorialId(Long historialId);
+public interface TransactionRepository extends JpaRepository<Transaction, Integer>{
+	List<Transaction> findAllByFlowId(int flowId);
+	List<Transaction> findAllByHistorialId(int historialId);
 }
