+++ src/main/java/com/acme/yupanaapi/domain/service/FlowService.java
@@ -12,12 +12,12 @@ import com.acme.yupanaapi.resource.UserWalletResource;
 import org.springframework.http.ResponseEntity;
 
 public interface FlowService {
-    Flow getFlowById(Long flowId);
-    Flow createFlow(Flow flow, Long walletId, Long sellerId);
-    Flow updateFlow(Long flowId,Long walletId, Long sellerId, Flow flowRequest);
-    ResponseEntity<?> deleteFlow(Long flowId);
-    List<Flow> getAllByWalletId(Long walletId);
-    List<Flow> getAllByWalletIdAndDeadlineDate(Long walletId, Date date);
-    Flow getLastFlow(Long walletId);
-    UserWalletResource getData(Long walletId);
+    Flow getFlowById(int flowId);
+    Flow createFlow(Flow flow, int walletId, int sellerId);
+    Flow updateFlow(int flowId,int walletId, int sellerId, Flow flowRequest);
+    ResponseEntity<?> deleteFlow(int flowId);
+    List<Flow> getAllByWalletId(int walletId);
+    List<Flow> getAllByWalletIdAndDeadlineDate(int walletId, Date date);
+    Flow getLastFlow(int walletId);
+    UserWalletResource getData(int walletId);
 }
