+++ src/main/java/com/acme/yupanaapi/domain/service/SellerService.java
@@ -4,11 +4,11 @@ import com.acme.yupanaapi.domain.model.Seller;
 import org.springframework.http.ResponseEntity;
 
 public interface SellerService {
-    Seller getSellerByIdAndUserId(Long sellerId, Long userId);
-    Seller getSellerById(Long sellerId);
+    Seller getSellerByIdAndUserId(int sellerId, int userId);
+    Seller getSellerById(int sellerId);
     Seller getSellerByBusinessName(String businessName);
-    Seller createSeller(Seller seller, Long userId);
-    Seller updateSeller(Long sellerId, Long userId, Seller sellerRequest);
-    ResponseEntity<?> deleteSeller(Long sellerId);
+    Seller createSeller(Seller seller, int userId);
+    Seller updateSeller(int sellerId, int userId, Seller sellerRequest);
+    ResponseEntity<?> deleteSeller(int sellerId);
 }
