+++ src/main/java/com/acme/yupanaapi/domain/service/SubscriptionService.java
@@ -5,9 +5,9 @@ import com.acme.yupanaapi.domain.model.Subscription;
 import java.util.List;
 
 public interface SubscriptionService {
-	Subscription createSubscription(Subscription subscription, Long walletId);
-	Subscription updateSubscription(Subscription subscriptionEntity, Long subscriptionId, Long walletId);
-	ResponseEntity<?> deleteSubscription(Long subscriptionId);
-	Subscription getSubscriptionById(Long subscriptionId);
-	List<Subscription> getAllByWalletId(Long walletId);
+	Subscription createSubscription(Subscription subscription, int walletId);
+	Subscription updateSubscription(Subscription subscriptionEntity, int subscriptionId, int walletId);
+	ResponseEntity<?> deleteSubscription(int subscriptionId);
+	Subscription getSubscriptionById(int subscriptionId);
+	List<Subscription> getAllByWalletId(int walletId);
 }
