+++ src/main/java/com/acme/yupanaapi/domain/service/WalletService.java
@@ -9,10 +9,10 @@ import java.util.List;
 
 
 public interface WalletService {
-	Wallet createWallet(Wallet entity, Long sellerId, Long userId);
-	Wallet updateWallet(Wallet entity, Long walletId, Long sellerId);
-	ResponseEntity<?> deleteWallet(Long walletId);
-	Wallet getWalletById(Long walletId);
-	Wallet getWalletByIdAndUserId(Long walletId, Long userId);
-	List<Wallet> getAllBySellerId(Long sellerId);
+	Wallet createWallet(Wallet entity, int sellerId, int userId);
+	Wallet updateWallet(Wallet entity, int walletId, int sellerId);
+	ResponseEntity<?> deleteWallet(int walletId);
+	Wallet getWalletById(int walletId);
+	Wallet getWalletByIdAndUserId(int walletId, int userId);
+	List<Wallet> getAllBySellerId(int sellerId);
 }
