+++ src/main/java/com/acme/yupanaapi/resource/TransactionResource.java
@@ -3,15 +3,12 @@ package com.acme.yupanaapi.resource;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-import javax.validation.constraints.NotNull;
 import java.util.Date;
 
 @Getter
 @Setter
 public class TransactionResource {
-    private Long id;
+    private int id;
     private String transactionName;
     private Date transactionDate;
     private Float amount;
