+++ src/main/java/com/acme/yupanaapi/resource/UserResource.java
@@ -3,13 +3,11 @@ package com.acme.yupanaapi.resource;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Lob;
-import javax.validation.constraints.NotNull;
 
 @Getter
 @Setter
 public class UserResource {
-    private Long id;
+    private int id;
     private String documentType;
     private String documentNumber;
     private String name;
